Seat information for confirmation emails. Issue #14734
We just needed to get the seatInfo variable in the right hook in order for it
to be available to the email templates. This also required a modification to
CiviCRM core to get participantIDs available in the email template.

Also moved the javascript around so we can have it in it's own file making it
easier to edit.