Fix problem with insertBefore duplication. Issue #14735.
insertBefore is supposed to remove the element it is inserting from the DOM
and insert it into its new place, but it doesn't seem to be working that way,
so I added some code to delete the table I was using to hold the section I'm
inserting. Without this, the form fields in my section were getting sent
twice with different values.