Implemented the allowed subscriptien select. Issue #14735.
Implemented the "Allowed Subscription" field. I wanted to do this in the
postProcess hook for the ManageEvent form, but for some weird reason that
hook isn't getting called when an event gets created for the first time. So I
put my code for this in the post hook for when events get saved or updated
like Jim did for his other fields. I think it would probably be better to
figure out why that hook isn't getting called by CiviCRM, but I thought that
might take awhile and I don't want to spend the time right now.