All the way through registration for subscription. Issue #14735.
Got all the way through registration for subscription, but bypassed quite a
few issues. If you enter an email for a Flex Pass, it bypasses all the
payment checks. Also, it isn't checking the values input for the ticket
fields. It really should just ignore what comes from the web and force it to
the subscription numbers.

Fixed up some of the javascript issues with the total fee, but it doesn't
survive reloads and probably not back either.