Get seat map to show up on backend.
The backend can use the same javascript file as the front end now, so getting
rid of the javascript in the template fixes it so the seatmap shows up again
on the edit participant page. I don't think the seatmap is smart, though, so
it is dangerous to use it on the back end because it doesn't deselect the
seats that the person used.