Implemented subscription selection. Issue #15138.
Had to rework things a bit because the earlier code made the assumption that
you could only use one subscription per email address.

Also reworked the Javascript code to make in object oriented because it was
getting a bit unwieldly as functions and some of the code now needed to keep
some state information about the subscription data.