Fixed problem with price fields and multiple allowed subscriptions.
If an event allowed multiple subscriptions then you might get 0s in the
price fields because it was using the wrong subscription for the price
fields. Now I'm sending the subscription_participant_id along with the
price fields and it matches that with the chosen subscription to fill in
the price fields correctly.