Fixed losing data with back button. Issue #15194.
When the user clicks on 'Go Back' or clicks on the back button in the browser
you'd like them to go back to a page that looks just like the one they filled
out. This stores the email address so it can put it back in when you go back
to the first page.