Fixed some more submission corner cases.
Always clear all the seats for both general and subscription categories when
showing the register page because we don't know if this use has already
selected one or the other.

Reload the general admission seatmap if we get an error when looking up a
subscription so the user can buy tickets normally.