Fixed price fields that don't have an association. Issue #14735
I had some checks to help get event price fields setup by printing out errors
if it can't find all the associations, but we actually want it to work such
that if the event doesn't have all the price fields it still works with the
ones that are there and also the subscription hides any price fields it
doesn't know about. So this fixes the problem Mick mentioned in comment #3.