Fixed logged in registration and some particpant params. Issue #15242. Issue #15236.
If you were logged in and you tried to do a registration using flex pass, the
flex pass would get assigned to the person you were logged in as instead of
the person associated with the flex pass email address you input. I've kind
of hacked this in using the registeredById field which is normally used for
the case where one person is registering another person for an event.