Fixed handling of radio button field disabling.
If there is a required radio button field in the price set, we need to
select one of its options. To make things easier for now, we are assuming
that any radio button is requied and has one choice with the data-amount
attribute set to 0. We select that one for submission.

Also found a bug where we were setting the value field of radio buttons to
'' when clearing the form. We shouldn't mess with radio field values because
it will submit '' when selected if we do that.