Fixed incorrect clearing of allowed subscriptions. Issue #15294.
Because we were deep in the database event creation/editing hooks it was hard
to tell whether or not the allowed subscription field was in the form or not
making it hard to differentiate a blank entry from being on a different page.
I've fixed this by fixing the problem that caused us to use this hook in the
first place. I patched CiviCRM core to fire the postProcess hook for
CRM_Event_Form_ManageEvent_EventInfo when the event gets created. I've
filed this with CiviCRM as CRM-13221.