Fixed continue,back,continue no seats problem. Issue #15329.
The seatmap stuff was only happening on pages with _POST, but when you are
getting the error version of the Register form it uses a _GET. I think the
snippet stuff I added accomplishes the same thing as what the original _POST
stuff was for.

I also adjusted the clearing fields thing slightly so that it doesn't clear
the fields on page load. Now it remembers what was in that field and only
clears the fields if the current value is different then the original value.
I think that's what most people would expect.