Let users change seats on back end. Issue #15218.
Sometimes they need to change the seats on the back end to correct mistakes
or just to be nice to the customers.

With the rearrangement of the code, the code that makes the payment fields
not required for subscription users wasn't getting called because that code
moved to after a POST/_snippet check. So I've moved it back up before that
check so that it happens at the right times.

The code that clears out the previously selected seats was passing in the
place_map_category to fill in some ids for which seats to clear, but I think
we should just get this from the seat data in the database itself. So I added
the needed columns to the returned columns in Seat::loadForSid.