Fixed seat choices display for flex passes on back-end. Issue #15430.
It looks like if there was no money associated with a participant
record, Civi let's you edit the price sets which results in a different
layout where the class is missing that we use to figure out where to put
the seat information. So I added a third fallback that works for these pages.