Removed seat limit. Issue #15416.
Nina says they frequently allow groups larger than 15, so they don't want a
limit on seats. I actually tried setting the seatlimit to -1 which the code
supports, but if you do that it breaks the javascript code in
js/civiboxoffice.js which checks that the number of tickets and seats
selected match. Instead I am setting the limit to 1000 which is bigger than
any venue they have.