Now seatmap works with CiviDiscount. Issue #15369.
They want to be able to use CiviDiscount with events with seatmaps, but it
wasn't working because the seatmap code adds a submit handler that checks to
see if the number of tickets bought matches the number of seats selected, but
the CiviDiscount lookup tries to do a whole page submit before they choose
seats which causes that validator to show an error.

What I've done here is to try and detect when they are doing the "Apply"
button for CiviDiscount instead of the "Continue>>" button down at the
bottom. If they are doing the "Apply" button, then we skip the seat
validation check.