Implemented GUI setting of price set associations. Issue #15206.
When people who have a subscription go to use the subscription, we need to
map the categories of tickets bought with the subscription to the categories
of tickets for the peformance. For example, the subscription might use a
price set with 'Adult' and 'Child' and the subscriber may have bought 1 Adult
and 2 Child. Now the performance may have a price set with categories like
'Adults' and 'Children' so we need to know that we need to map the 1 Adult to
Adults and the 2 Child to Chidren. This implements a GUI interface on the
Manage Event - Fees page to set those rules up.