Only show seat map warning if event has a seat map. Issue #18173.
BACT asked that we show a message saying that if the seat map doesn't
appear, the user should try another web browser. This was showing up on
every event including ones that didn't have seat maps because I knew it
would be a tricky to get this message to display on browsers which we know
are having some sort of javascript problem.

I eventually decided to put it in the main Register template because that's
the only way I can get it to show up in the middle of the event
registration page without using javasacript (that's in a separate commit in
their Drupal repository).

I had to also change the code that does the buildForm part of the seatMap
to not exit early on the main page. Normally we want all that to happen in
a snippet, but we need the code that decides whether or not this event has
a seat map on the non-snippet load of this page in order to set the new
hasSeatMap template var.